fixup! During rendering, allow yielded block to contribute references
