+++ lib/view_model/active_record/controller.rb
@@ -3,7 +3,6 @@
 require 'view_model/active_record/controller_base'
 require 'view_model/active_record/collection_nested_controller'
 require 'view_model/active_record/singular_nested_controller'
-require 'view_model/controller/migration_versions'
 
 # Controller for accessing an ViewModel::ActiveRecord
 # Provides for the following routes:
@@ -17,7 +16,6 @@ module ViewModel::ActiveRecord::Controller
   include ViewModel::ActiveRecord::ControllerBase
   include ViewModel::ActiveRecord::CollectionNestedController
   include ViewModel::ActiveRecord::SingularNestedController
-  include ViewModel::Controller::MigrationVersions
 
   def show(scope: nil, viewmodel_class: self.viewmodel_class, serialize_context: new_serialize_context(viewmodel_class: viewmodel_class))
     view = nil
@@ -68,32 +66,6 @@ module ViewModel::ActiveRecord::Controller
     render_viewmodel(nil)
   end
 
-  included do
-    etag { migrated_deep_schema_version }
-  end
-
-  def parse_viewmodel_updates
-    super.tap do |update_hash, refs|
-      if migration_versions.present?
-        migrator = ViewModel::UpMigrator.new(migration_versions)
-        migrator.migrate!({ 'data' => update_hash, 'references' => refs })
-      end
-    end
-  end
-
-  def prerender_viewmodel(...)
-    super do |jbuilder|
-      yield(jbuilder) if block_given?
-
-      # migrate the resulting structure before it's serialized to a json string
-      if migration_versions.present?
-        tree = jbuilder.attributes!
-        migrator = ViewModel::DownMigrator.new(migration_versions)
-        migrator.migrate!(tree)
-      end
-    end
-  end
-
   private
 
   # Viewmodel ids are permitted to be either integers or strings
@@ -118,9 +90,5 @@ module ViewModel::ActiveRecord::Controller
 
   def viewmodel_id
     parse_param(:id, with: ViewmodelIdSerializer)
-  end
-
-  def migrated_deep_schema_version
-    ViewModel::Migrator.migrated_deep_schema_version(viewmodel_class, migration_versions, include_referenced: true)
   end
 end
